+++ src/components/Book.js
@@ -1,6 +1,6 @@
 import { PropTypes } from 'prop-types';
 import { useDispatch } from 'react-redux';
-import { removeBook } from 'redux/books/booksSlice';
+import { removeBook } from 'redux/books/BooksThunk';
 
 const Book = ({ id, title, author }) => {
   const dispatch = useDispatch();
